requirements.txt
bs4=4.10.0
pandas=1.3.4
pillow=8.4.0
requests=2.26.0